refactored the activity feed to children components
